// src/types/home.d.ts
export interface HomeProductCategories {
  id: string;
  name: string;
  image: string;
}

export interface HomeNewArrivals {
  name: string;
  code: string;
  price: string;
  image: string;
}

export interface HomeTopProducts {
  name: string;
  code: string;
  price: string;
  image: string;
}

export interface HomeSuitableFor {
  Icon: any;
  name: string;
}